Delete OCWSerializer and modify ocw tests
